backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
cryptography==41.0.7
pydantic==2.5.0
asyncio==3.4.3
apscheduler==3.10.4
pandas==2.1.4
numpy==1.24.3
# Add these when you're ready for ML
# torch==2.1.0
# transformers==4.35.2
# scikit-learn==1.3.2